requirements.txt
aiofiles==0.8.0
aiohttp==3.8.1
aiosignal==1.2.0
aiosqlite==0.17.0
APScheduler==3.8.1
async-timeout==4.0.2
attrs==21.4.0
black==21.12b0
cachetools==5.0.0
charset-normalizer==2.0.10
click==8.0.3
colorlog==6.6.0
d20==1.1.2
frozenlist==1.3.0
hikari==2.0.0.dev105
hikari-lightbulb==2.1.3
idna==3.3
lark-parser==0.9.0
multidict==5.2.0
mypy-extensions==0.4.3
pathspec==0.9.0
platformdirs==2.4.1
pure25519==0.0.1
python-dotenv==0.19.2
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
six==1.16.0
tomli==1.2.3
typing-extensions==4.0.1
tzdata==2021.5
tzlocal==4.1
uvloop==0.16.0
yarl==1.7.2
